Version 0.6.5   24-jan-2011 -------------  * Upgraded libpng to version 1.4.5-optipng [private]  * Upgraded zlib to version 1.2.5-optipng [private] !! Fixed the I/O states (in libpng 1.4.5);    they caused incorrect file reads in some rare cases.    (Thanks to [M*A*S*H] and Dmitry Marakasov for the report.) !! Fixed processing of PNG files with chunks of size 0.    (Thanks to Matthew Fearnley for the report.)  ! Fixed a display error in the TIFF import.    (Thanks to Piotr Bandurski for the fix.)  + Improved checking of the arguments of -f, -zc, -zm and -zs.  - Removed quirks from the rangeset option argument syntax.  ! Fixed a build issue under the system-supplied libpng-1.4.    (Thanks to Petr Gajdos for the fix.)  * Resolved forward-compatibility issues regarding libpng-1.5;    however, the system-supplied libpng-1.5 is not yet supported.  * Added various enhancements to the configure+make build system.    (Thanks in part to Elias Pipping and Dmitri Zubko.)

+++ lib/pngxtern/pngxrbmp.c
@@ -322,7 +322,7 @@ bmp_read_rows(png_bytepp begin_row, png_bytepp end_row, size_t row_size,
             }
             else  /* b2 >= 3 bytes in absolute mode */
             {
-               len = (b2 <= endn - crtn) ? b2 : (endn - crtn);
+               len = (b2 <= endn - crtn) ? b2 : (unsigned int)(endn - crtn);
                if (bmp_fread_fn(*crt_row, crtn, len, stream) != len)
                   break;
                crtn += len;
@@ -330,7 +330,7 @@ bmp_read_rows(png_bytepp begin_row, png_bytepp end_row, size_t row_size,
          }
          else  /* b1 > 0 bytes in run-length encoded mode */
          {
-            len = (b1 <= endn - crtn) ? b1 : (endn - crtn);
+            len = (b1 <= endn - crtn) ? b1 : (unsigned int)(endn - crtn);
             bmp_memset_fn(*crt_row, crtn, (int)b2, len);
             crtn += len;
          }
